task1_1/requirements.txt
certifi==2019.11.28
chardet==3.0.4
facenet-pytorch==2.0.1
idna==2.8
numpy==1.18.1
pandas==0.25.3
Pillow==7.0.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.14.0
tqdm==4.42.0
urllib3==1.25.8
wincertstore==0.2